+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Provisioner will link the Provisioned Certificate to the corresponed CO Person Role(COU)
+- Pick the Certificate linked under the delete CO Person Role to use for user (de)provisioning
 
 
 ## [0.2.0] - 2020-10-14